// src/api/generated/models/paymentStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const PaymentStatus = {
  PENDING: 'PENDING',
  UPLOADED: 'UPLOADED',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
} as const;