// src/api/generated/models/shippingInvoiceStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type ShippingInvoiceStatus = typeof ShippingInvoiceStatus[keyof typeof ShippingInvoiceStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ShippingInvoiceStatus = {
  DRAFT: 'DRAFT',
  ISSUED: 'ISSUED',
  IN_TRANSIT: 'IN_TRANSIT',
  DELIVERED: 'DELIVERED',
} as const;