// src/api/generated/models/supplierQuoteStatus.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type SupplierQuoteStatus = typeof SupplierQuoteStatus[keyof typeof SupplierQuoteStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SupplierQuoteStatus = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  PARTIALLY_ACCEPTED: 'PARTIALLY_ACCEPTED',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  EXPIRED: 'EXPIRED',
} as const;